Cached whether Wyckoff positions are unique or not
Cached whether Wyckoff positions are unique or not in wyckoffDatabase.h. It is added as a fourth element to the tuple. Now, SpgGen::containsUniquePosition() just checks the tuple and returns the value from it. This should theoretically speed up the program a lot since 36.02% of the program was spent calculating whether the Wyckoff positions are unique or not.